+++ packages/wallet-ts/src/Cosmos/index.ts
@@ -1,4 +1,5 @@
 export * from './CosmosQuery'
 export * from './CosmosWallet'
+export * from './CosmosClient'
 export * from './utils'
 export * from './types'