+++ packages/wallet-ts/src/Keplr/index.ts
@@ -1,3 +1,2 @@
 export * from './KeplrWallet'
 export * from './utils'
-export * from './types'