+++ packages/wallet-ts/src/wallet-strategy/strategies/Keplr.ts
@@ -3,20 +3,16 @@ import {
   AccountAddress,
   ChainId,
   EthereumChainId,
+  CosmosChainId,
 } from '@injectivelabs/ts-types'
 import { Web3Exception } from '@injectivelabs/exceptions'
-import {
-  BigNumberInBase,
-  DEFAULT_GAS_LIMIT,
-  DEFAULT_GAS_PRICE,
-} from '@injectivelabs/utils'
+import { DEFAULT_STD_FEE } from '@injectivelabs/utils'
 import type Web3 from 'web3'
+import { createTxRaw } from '@injectivelabs/tx-ts'
 import { KeplrWallet } from '../../keplr'
 import { ConcreteWalletStrategy } from '../types'
 import BaseConcreteStrategy from './Base'
 import { CosmosWallet } from '../../cosmos/CosmosWallet'
-import { CosmosChainId } from '../../keplr/types'
-import { createSignedTx } from '../../transaction'
 
 export default class Keplr
   extends BaseConcreteStrategy
@@ -87,7 +83,7 @@ export default class Keplr
     _options: { address: AccountAddress; chainId: ChainId },
   ): Promise<string> {
     const { keplrWallet } = this
-    const txRaw = createSignedTx(signResponse)
+    const txRaw = createTxRaw(signResponse)
 
     try {
       return keplrWallet.broadcastTx(txRaw)
@@ -114,25 +110,13 @@ export default class Keplr
       signer,
     })
 
-    const fee = {
-      amount: [
-        {
-          amount: new BigNumberInBase(DEFAULT_GAS_LIMIT)
-            .times(DEFAULT_GAS_PRICE)
-            .toString(),
-          denom: 'inj',
-        },
-      ],
-      gas: DEFAULT_GAS_LIMIT.toString(),
-    }
-
     return cosmWallet.signTransaction({
+      address,
       chainId,
+      fee: DEFAULT_STD_FEE,
       message: transaction.message,
       memo: transaction.memo,
       pubKey: Buffer.from(key.pubKey).toString('base64'),
-      address,
-      fee,
     })
   }