+++ packages/wallet-ts/src/wallet-strategy/strategies/Trezor/index.ts
@@ -174,7 +174,7 @@ export default class Trezor
       gasLimit: addHexPrefix(txData.gas),
       maxFeePerGas: addHexPrefix(txData.gasPrice || txData.maxFeePerGas),
       maxPriorityFeePerGas: addHexPrefix(
-        txData.maxPriorityFeePerGas || '0xB2D05E00' /* 3 Gwei in HEX */,
+        txData.maxPriorityFeePerGas || '0x77359400' /* 2 Gwei in HEX */,
       ),
     }
     const tx = FeeMarketEIP1559Transaction.fromTxData(eip1559TxData, {