+++ deprecated/dmm-consumer/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.5.9](https://github.com/InjectiveLabs/injective-ts/compare/@injectivelabs/dmm-consumer@0.5.8...@injectivelabs/dmm-consumer@0.5.9) (2022-06-10)
+
+
+### Features
+
+* added tx-ts ([ae9ebc7](https://github.com/InjectiveLabs/injective-ts/commit/ae9ebc7e2c34eaf60e894cd70e6b0778e6b71bbf))
+
+
+
+
+
 ## [0.5.8](https://github.com/InjectiveLabs/injective-ts/compare/@injectivelabs/dmm-consumer@0.5.7...@injectivelabs/dmm-consumer@0.5.8) (2022-06-08)
 
 **Note:** Version bump only for package @injectivelabs/dmm-consumer